Fixed BigDecimal in firestore. Refactored all repositories. Modified some views. Typo of Products into Product collection but it is okay.

+++ app/src/main/java/com/teampym/onlineclothingshopapplication/data/db/UserInformationDao.kt
@@ -18,7 +18,7 @@ interface UserInformationDao {
     suspend fun getCurrentUser(userId: String): UserInformation
 
     @Query("UPDATE table_users SET totalOfCart = :totalOfCart WHERE userId = :userId")
-    suspend fun updateTotalOfCart(userId: String, totalOfCart: BigDecimal)
+    suspend fun updateTotalOfCart(userId: String, totalOfCart: Double)
 
     @Update
     suspend fun update(userInformation: UserInformation)
